requirements.txt
requests>=2.25.0
pandas>=1.1.4
tqdm>=4.54.1
eth-brownie>=1.12.2